requirements.txt
tensorflow>=2.10.0
matplotlib
numpy
opencv-python
Pillow
streamlit
